chore: getRangeLabel 분리
